logs interface
