+++ internal/querylog/jsonfile/qlogfile_test.go
@@ -1,4 +1,4 @@
-package querylog
+package jsonfile
 
 import (
 	"encoding/binary"
