+++ internal/querylog/jsonfile/qlogreader_test.go
@@ -1,4 +1,4 @@
-package querylog
+package jsonfile
 
 import (
 	"io"
